Create separate activity for PoC
